Rename directory "pq" as "quantization"
