+++ jvector-base/src/main/java/io/github/jbellis/jvector/util/PhysicalCoreExecutor.java
@@ -16,7 +16,7 @@
 package io.github.jbellis.jvector.util;
 
 import io.github.jbellis.jvector.graph.GraphIndexBuilder;
-import io.github.jbellis.jvector.pq.ProductQuantization;
+import io.github.jbellis.jvector.quantization.ProductQuantization;
 
 import java.io.Closeable;
 import java.util.concurrent.ForkJoinPool;
