+++ jvector-examples/src/main/java/io/github/jbellis/jvector/example/distancesNVQ.java
@@ -18,8 +18,8 @@ package io.github.jbellis.jvector.example;
 
 import io.github.jbellis.jvector.example.util.SiftLoader;
 import io.github.jbellis.jvector.graph.ListRandomAccessVectorValues;
-import io.github.jbellis.jvector.pq.NVQVectors;
-import io.github.jbellis.jvector.pq.NVQuantization;
+import io.github.jbellis.jvector.quantization.NVQVectors;
+import io.github.jbellis.jvector.quantization.NVQuantization;
 import io.github.jbellis.jvector.vector.VectorSimilarityFunction;
 import io.github.jbellis.jvector.vector.VectorUtil;
 import io.github.jbellis.jvector.vector.types.VectorFloat;
